Prevent error due to missing wazuh-control from previous installation

+++ packages/rpms/SPECS/wazuh-agent.spec
@@ -197,9 +197,14 @@ if ! getent passwd wazuh > /dev/null 2>&1; then
   useradd -g wazuh -G wazuh -d %{_localstatedir} -r -s /sbin/nologin wazuh
 fi
 
-# Get version information
-OLD_VERSION=`%{_localstatedir}/bin/wazuh-control info -v`
-MAJOR=$(echo $OLD_VERSION | cut -dv -f2 | cut -d. -f1)
+# Get version information (only for old installations)
+if [ -f "%{_localstatedir}/bin/wazuh-control" ]; then
+  OLD_VERSION=`%{_localstatedir}/bin/wazuh-control info -v 2>/dev/null`
+  MAJOR=$(echo $OLD_VERSION | cut -dv -f2 | cut -d. -f1)
+elif [ -f %{_localstatedir}/VERSION.json ]; then
+  OLD_VERSION=`grep -oP '(?<="version": ")[^"]*' %{_localstatedir}/VERSION.json 2>/dev/null`
+  MAJOR=$(echo $OLD_VERSION | cut -d. -f1)
+fi
 
 # Stop the services to upgrade the package
 if [ $1 = 2 ]; then
@@ -224,7 +229,7 @@ if [ $1 = 2 ]; then
 fi
 
 # Remove old databases if upgrading from pre 5.X to 5.X
-if [ $MAJOR -lt 5 ]; then
+if [ -n "$MAJOR" ] && [ "$MAJOR" -lt 5 ]; then
   if [ -f %{_localstatedir}/queue/syscollector/db/local.db ]; then
     rm -f %{_localstatedir}/queue/syscollector/db/local.db
   fi
